Bring back main dashboard background color
We removed the background to the main background color in the commit
below. The problem we had was that while using 2 or even 3 columns,
blank "tiles" would get color in with the background color and so giving
us a janky look. The problem we have now is that some themers would like
to make use of this background color to give their themes a certain look.

We're setting it to transparent so it leaves us with the same look we have
now and opening that door again for themers to use this background color.

https://github.com/DirtyUnicorns/android_packages_apps_Settings/commit/513a19073556744dc68cb1a63201babab7d82f24

Change-Id: I40ff32c8465459315ee61e1244cf311ebe85b51a

Conflicts:
	res/values/projekt_colors.xml