Add Ambient display category / Preference controllers to handle switches
This commit puts both switches inside a category 'Ambient options' while preserving
the conditions originally placed on to them by Google. It also continues to give us
the option of removing the entire category for devices such as the OP3/OP5 and any
device that uses a custom Doze package.

- DozePreferenceController was used as a template for DozeAlwaysOn and DozeCategory

- AlwaysOn 'isAvailable' code taken from the commit below
https://github.com/DirtyUnicorns/android_frameworks_base/blob/d7e5a2b886225945b85de365aab1668d558fe0a1/core/java/com/android/internal/hardware/AmbientDisplayConfiguration.java#L82-L85

- Add preference and controller for custom doze package (com.custom.ambient.display)

Change-Id: If25aa1cba5156b02d37b8754fede30bc87742be8